refactor: [#1211] move test to torrust_tracker_primitives::core
