+++ packages/primitives/src/core.rs
@@ -56,3 +56,26 @@ impl ScrapeData {
         self.files.insert(*info_hash, SwarmMetadata::zeroed());
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use bittorrent_primitives::info_hash::InfoHash;
+
+    use crate::core::ScrapeData;
+
+    fn sample_info_hash() -> InfoHash {
+        "3b245504cf5f11bbdbe1201cea6a6bf45aee1bc0".parse::<InfoHash>().unwrap()
+    }
+
+    #[test]
+    fn it_should_be_able_to_build_a_zeroed_scrape_data_for_a_list_of_info_hashes() {
+        // Zeroed scrape data is used when the authentication for the scrape request fails.
+
+        let sample_info_hash = sample_info_hash();
+
+        let mut expected_scrape_data = ScrapeData::empty();
+        expected_scrape_data.add_file_with_zeroed_metadata(&sample_info_hash);
+
+        assert_eq!(ScrapeData::zeroed(&vec![sample_info_hash]), expected_scrape_data);
+    }
+}
